refactor(VideoAssetCard): large/small variant names
