+++ src/script/components/MessagesList/Message/ContentMessage/asset/MultipartAssets/VideoAssetCard/VideoAssetSmall/VideoAssetSmall.styles.ts
